Creat Custom Snackbar and save jwt token on session
CustomSnackbar created and added to signin component in case of failed login.

SignIn component now sends an axios call to the login route on the backend. In success situations, it saves the user email and jwt token on session and redirects to home page. In case failure the snackbar is shown with the warning variant, telling the user that either the username or the password are incorrect