Added SignUp component and remov claim for phone number
Removed PhoneNumber claim from jwt Generate function, as it was causing a bug when the user registered without setting up a phone number.
Added SignUp component that calls the register route. In case of success it also takes care of the login part, adding on session the jwt and the email. In case of error, it shows a custom error Snackbar.