Start work on account page and partialy change authorization on adress controller
Allowed Anonymous on routes on Address for ease of testing on frontend side. Will change it back in the next merges.
Added account form page  that contains user info in a formik form. Will change design and integrate with material ui.
A base error page was created that just contains a loading circle.