Add logout functionality and component
Logout component created. Component redirects to logout route and opens a modal dialog giving the user the ability to decide if he wants to logout or cancel.
Upon logout, the session items jwtBearer and userEmail are deleated and the user is redirected to the homepage.
If the user cancels or clicks out of the dialog box, he gets redirected to the previous route that he was on.