Change Order Entity, OrderModel and order controller to connect user ID and address to order
Changed route in order controller on post to facilitate sending the userEmail, as to then get using it the user Id and add it to the mapping between OrderModel and Order. The option to add the user ID to the OrderModel was decided against, as on the frontend the user ID should be avoided, but accest to email is facile.

Added in Order an Address id as to take care of any possible ambiguity and facilitate code comprehention.

Added Address id in OrderModel as to get it directly from the frontend, as access to it is easily obtained after selecting the address.