Singup redirects to OrderComplete that takes order id via useParams from the route
Added "/order-complete/:orderId" route in App.tsx that will render OrderComplete component.
The component takes the orderId using the useParams hook.
The user gets redirected to the OrderComplete page after an order is submitted. User should also be able to get to this page if he wants to see any of his orders.